Refactored marketplace view
